refactor: bottom-sheet 폴더&파일 이름 변경 #128
